+++ Implementation/ADC_ADS1115_driver/ESP32_ESP_IDF/ADC_ADS1115_driver.h
@@ -24,7 +24,7 @@ public:
 
     bool is_conversion_complete() override;
     
-    float get_measured_voltage() override;
+    double get_measured_voltage() override;
 
     void set_samples_per_second(uint16_t rate) override;
